refac(91026): Extrai task gerar_xlsx_extrato_dres_async
