+++ sme_ptrf_apps/core/tasks/__init__.py
@@ -11,6 +11,7 @@ from .gerar_notificacao_proximidade_inicio_pc import gerar_notificacao_proximida
 from .gerar_notificacao_proximidade_fim_periodo_pc import gerar_notificacao_proximidade_fim_periodo_prestacao_conta_async
 from .gerar_notificacao_atraso_entrega_ajustes_pc import gerar_notificacao_atraso_entrega_ajustes_prestacao_de_contas_async
 from .gerar_notificacao_proximidade_fim_prazo_ajustes_pc import gerar_notificacao_proximidade_fim_prazo_ajustes_prestacao_conta_async
+from .gerar_xlsx_extrato_dres import gerar_xlsx_extrato_dres_async
